some stuffs

+++ src/entity/Entity.java
@@ -11,7 +11,7 @@ public class Entity extends ImageView {
 	public List<Image> up, upl, down, downl, left, right, nomovel, nomover;
 	public String direction;
 
-	public int spriteCounter = 0;
+	public long spriteCounter = 0;
 	public int spriteNum = 0;
 
 }
